+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Removed `gpu.js` from optional dependencies and marked config options `gpujs` and `gpuMode` as deprecated.
+
 ## [1.1.0] - 2021-08-12
 
 ### Changed